Move `PixelBufferCache` to separate file
