+++ ios/Frame Processor/FrameHostObject.mm
@@ -15,33 +15,6 @@
 
 namespace vision {
 
-PixelBufferCache::~PixelBufferCache() {
-  if (pixelBuffer != nil) {
-    CFRelease(pixelBuffer);
-  }
-}
-
-uint8_t* PixelBufferCache::getPixelBuffer() {
-  if (pixelBuffer == nil) {
-    auto imageBuffer = CMSampleBufferGetImageBuffer(frame.buffer);
-    CVPixelBufferLockBaseAddress(imageBuffer, 0);
-    void* buffer = CVPixelBufferGetBaseAddress(imageBuffer);
-    CVPixelBufferUnlockBaseAddress(imageBuffer, 0);
-    pixelBuffer = (uint8_t*)buffer;
-  }
-  return pixelBuffer;
-}
-
-size_t PixelBufferCache::getPixelBufferSize() {
-  if (pixelBufferSize == -1) {
-    auto imageBuffer = CMSampleBufferGetImageBuffer(frame.buffer);
-    size_t bytesPerRow = CVPixelBufferGetBytesPerRow(imageBuffer);
-    size_t height = CVPixelBufferGetHeight(imageBuffer);
-    pixelBufferSize = bytesPerRow * height;
-  }
-  return pixelBufferSize;
-}
-
 std::vector<jsi::PropNameID> FrameHostObject::getPropertyNames(jsi::Runtime& rt) {
   std::vector<jsi::PropNameID> result;
   result.push_back(jsi::PropNameID::forUtf8(rt, std::string("toString")));
